cub rework

+++ srcs/utils/utils3.c
@@ -6,7 +6,7 @@
 /*   By: wouhliss <wouhliss@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/23 18:23:37 by ybelatar          #+#    #+#             */
-/*   Updated: 2024/03/27 13:42:35 by wouhliss         ###   ########.fr       */
+/*   Updated: 2024/04/04 02:20:38 by wouhliss         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -49,9 +49,9 @@ void	*ft_calloc(size_t nmemb, size_t size)
 	return (res);
 }
 
-inline void	ft_swapd(double *a, double *b)
+inline void	ft_swapd(float *a, float *b)
 {
-	double	c;
+	float	c;
 
 	c = *a;
 	*a = *b;
